+++ novawallet/Modules/Staking/StakingMain/Relaychain/ViewModel/StakingStateViewModelFactory.swift
@@ -255,7 +255,7 @@ extension StakingStateViewModelFactory: StakingStateVisitorProtocol {
             .unstake,
             .rewardDestination,
             .setupValidators,
-            .proxyAction(from: state.commonData.proxy),
+            .proxyAction(from: state.commonData.proxy, chain: chainAsset.chain),
             .controllerAccount
         ].compactMap { $0 }
 
@@ -320,7 +320,7 @@ extension StakingStateViewModelFactory: StakingStateVisitorProtocol {
             .rewardDestination,
             .pendingRewards,
             .changeValidators(count: state.nomination.targets.count),
-            .proxyAction(from: state.commonData.proxy),
+            .proxyAction(from: state.commonData.proxy, chain: chainAsset.chain),
             .controllerAccount
         ].compactMap { $0 }
 
@@ -385,7 +385,7 @@ extension StakingStateViewModelFactory: StakingStateVisitorProtocol {
             .rewardDestination,
             .pendingRewards,
             .yourValidator,
-            .proxyAction(from: state.commonData.proxy),
+            .proxyAction(from: state.commonData.proxy, chain: chainAsset.chain),
             .controllerAccount
         ].compactMap { $0 }
