+++ novawallet/Modules/Vote/Governance/Validating/GovernanceErrorPresentable.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-protocol GovernanceErrorPresentable: BaseErrorPresentable, DelegationErrorPresentable {
+protocol GovernanceErrorPresentable: BaseErrorPresentable {
     func presentNotEnoughTokensToVote(
         from view: ControllerBackedProtocol,
         available: String,
@@ -20,6 +20,11 @@ protocol GovernanceErrorPresentable: BaseErrorPresentable, DelegationErrorPresen
     func presentVotesMaximumNumberReached(
         from view: ControllerBackedProtocol,
         allowed: String,
+        locale: Locale?
+    )
+
+    func presentSelfDelegating(
+        from view: ControllerBackedProtocol,
         locale: Locale?
     )
 
@@ -79,6 +84,23 @@ extension GovernanceErrorPresentable where Self: AlertPresentable & ErrorPresent
     ) {
         let title = R.string.localizable.govMaxVotesReachedTitle(preferredLanguages: locale?.rLanguages)
         let message = R.string.localizable.govMaxVotesReachedMessage(allowed, preferredLanguages: locale?.rLanguages)
+
+        let close = R.string.localizable.commonClose(preferredLanguages: locale?.rLanguages)
+
+        present(message: message, title: title, closeAction: close, from: view)
+    }
+
+    func presentSelfDelegating(
+        from view: ControllerBackedProtocol,
+        locale: Locale?
+    ) {
+        let title = R.string.localizable.govAddDelegateSelfErrorTitle(
+            preferredLanguages: locale?.rLanguages
+        )
+
+        let message = R.string.localizable.govAddDelegateSelfErrorMessage(
+            preferredLanguages: locale?.rLanguages
+        )
 
         let close = R.string.localizable.commonClose(preferredLanguages: locale?.rLanguages)
