+++ novawallet/Modules/Vote/Governance/Validating/GovernanceValidatorFactory.swift
@@ -2,7 +2,7 @@ import Foundation
 import BigInt
 import SoraFoundation
 
-protocol GovernanceValidatorFactoryProtocol: BaseDataValidatingFactoryProtocol, DelegationValidatorFactoryProtocol {
+protocol GovernanceValidatorFactoryProtocol: BaseDataValidatingFactoryProtocol {
     func enoughTokensForVoting(
         _ assetBalance: AssetBalance?,
         votingAmount: BigUInt?,
@@ -35,6 +35,12 @@ protocol GovernanceValidatorFactoryProtocol: BaseDataValidatingFactoryProtocol, 
         locale: Locale?
     ) -> DataValidating
 
+    func notSelfDelegating(
+        selfId: AccountId?,
+        delegateId: AccountId?,
+        locale: Locale?
+    ) -> DataValidating
+
     func notVoting(
         _ accountVotingDistribution: ReferendumAccountVotingDistribution?,
         tracks: Set<TrackIdLocal>?,
@@ -53,8 +59,6 @@ final class GovernanceValidatorFactory {
     weak var view: ControllerBackedProtocol?
 
     var basePresentable: BaseErrorPresentable { presentable }
-    var delegationErrorPresentable: DelegationErrorPresentable { presentable }
-
     let assetBalanceFormatterFactory: AssetBalanceFormatterFactoryProtocol
     let quantityFormatter: LocalizableResource<NumberFormatter>
     let presentable: GovernanceErrorPresentable
@@ -233,6 +237,22 @@ extension GovernanceValidatorFactory: GovernanceValidatorFactoryProtocol {
             let numberOfVotes = accountVotingDistribution.votedTracks[track]?.count ?? 0
 
             return numberOfVotes < Int(accountVotingDistribution.maxVotesPerTrack)
+        })
+    }
+
+    func notSelfDelegating(
+        selfId: AccountId?,
+        delegateId: AccountId?,
+        locale: Locale?
+    ) -> DataValidating {
+        ErrorConditionViolation(onError: { [weak self] in
+            guard let view = self?.view else {
+                return
+            }
+
+            self?.presentable.presentSelfDelegating(from: view, locale: locale)
+        }, preservesCondition: {
+            selfId != nil && delegateId != nil && selfId != delegateId
         })
     }
